Store custom simplestreams key in /etc/juju/publicsimplestreamskey
