+++ cloudconfig/userdatacfg_unix.go
@@ -11,6 +11,7 @@ import (
 	"fmt"
 	"io/ioutil"
 	"path"
+	"path/filepath"
 	"strings"
 	"text/template"
 	"time"
@@ -22,10 +23,11 @@ import (
 	"github.com/juju/utils/proxy"
 	goyaml "gopkg.in/yaml.v1"
 
+	"github.com/juju/juju/agent"
 	"github.com/juju/juju/cloudconfig/cloudinit"
 	"github.com/juju/juju/environs/config"
 	"github.com/juju/juju/environs/imagemetadata"
-	"github.com/juju/juju/juju/osenv"
+	"github.com/juju/juju/environs/simplestreams"
 	"github.com/juju/juju/service"
 	"github.com/juju/juju/service/systemd"
 	"github.com/juju/juju/service/upstart"
@@ -205,15 +207,8 @@ func (w *unixConfigure) ConfigureJuju() error {
 	}
 
 	if w.icfg.PublicImageSigningKey != "" {
-		keyFile := "/home/ubuntu/simplestreamspublickey"
+		keyFile := filepath.Join(agent.DefaultConfDir, simplestreams.SimplestreamsPublicKeyFile)
 		w.conf.AddRunTextFile(keyFile, w.icfg.PublicImageSigningKey, 0644)
-		script := fmt.Sprintf(`([ ! -e /home/ubuntu/.profile ] || grep -q '%s' /home/ubuntu/.profile) || `+
-			`printf '\n#Added by juju\nexport %s=%s\n' >> /home/ubuntu/.profile`,
-			osenv.JujuImageStreamsPublicKeyFileEnvKey,
-			osenv.JujuImageStreamsPublicKeyFileEnvKey,
-			keyFile,
-		)
-		w.conf.AddScripts(script)
 	}
 
 	// Make the lock dir and change the ownership of the lock dir itself to
