+++ environs/bootstrap/bootstrap.go
@@ -5,6 +5,7 @@ package bootstrap
 
 import (
 	"fmt"
+	"io/ioutil"
 	"os"
 	"path/filepath"
 
@@ -168,7 +169,7 @@ func Bootstrap(ctx environs.BootstrapContext, environ environs.Environ, args Boo
 	}
 
 	ctx.Infof("Installing Juju agent on bootstrap instance")
-	publicKey, err := imagemetadata.UserPublicSigningKey()
+	publicKey, err := userPublicSigningKey()
 	if err != nil {
 		return err
 	}
@@ -183,6 +184,23 @@ func Bootstrap(ctx environs.BootstrapContext, environ environs.Environ, args Boo
 	}
 	ctx.Infof("Bootstrap agent installed")
 	return nil
+}
+
+func userPublicSigningKey() (string, error) {
+	signingKeyFile := os.Getenv("JUJU_STREAMS_PUBLICKEY_FILE")
+	signingKey := ""
+	if signingKeyFile != "" {
+		path, err := utils.NormalizePath(signingKeyFile)
+		if err != nil {
+			return "", errors.Annotatef(err, "cannot expand key file path: %s", signingKeyFile)
+		}
+		b, err := ioutil.ReadFile(path)
+		if err != nil {
+			return "", errors.Annotatef(err, "invalid public key file: %s", path)
+		}
+		signingKey = string(b)
+	}
+	return signingKey, nil
 }
 
 // setBootstrapTools returns the newest tools from the given tools list,
