+++ environs/tools/simplestreams_test.go
@@ -1018,7 +1018,7 @@ func (*metadataHelperSuite) TestReadMetadataPrefersNewIndex(c *gc.C) {
 }
 
 type signedSuite struct {
-	origKey string
+	coretesting.BaseSuite
 }
 
 var testRoundTripper *jujutest.ProxyRoundTripper
@@ -1055,16 +1055,17 @@ func (s *signedSuite) SetUpSuite(c *gc.C) {
 	imageData["/signed/streams/v1/tools_metadata.sjson"] = string(signedData)
 	testRoundTripper.Sub = jujutest.NewCannedRoundTripper(
 		imageData, map[string]int{"signedtest://unauth": http.StatusUnauthorized})
-	s.origKey = tools.SetSigningPublicKey(sstesting.SignedMetadataPublicKey)
+	s.PatchValue(&simplestreams.SimplestreamsJujuPublicKey, sstesting.SignedMetadataPublicKey)
 }
 
 func (s *signedSuite) TearDownSuite(c *gc.C) {
 	testRoundTripper.Sub = nil
-	tools.SetSigningPublicKey(s.origKey)
 }
 
 func (s *signedSuite) TestSignedToolsMetadata(c *gc.C) {
-	signedSource := simplestreams.NewURLDataSource("test", "signedtest://host/signed", utils.VerifySSLHostnames)
+	signedSource := simplestreams.NewURLSignedDataSource(
+		"test", "signedtest://host/signed", sstesting.SignedMetadataPrivateKey, utils.VerifySSLHostnames,
+	)
 	toolsConstraint := tools.NewVersionedToolsConstraint(version.MustParse("1.13.0"), simplestreams.LookupParams{
 		CloudSpec: simplestreams.CloudSpec{"us-east-1", "https://ec2.us-east-1.amazonaws.com"},
 		Series:    []string{"precise"},
